% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAST.R
\name{CAST.audit}
\alias{CAST.audit}
\title{Given audit data, compute p.values and all that.}
\usage{
CAST.audit(Z, audit = NULL, plan = NULL, ...)
}
\arguments{
\item{Z}{elec.data object (voter matrix)}

\item{audit}{A data.matrix holding the audit data, if the Z object
does not have one, or if it is desirable to override it.  If both
the Z object has an audit object and audit is not null, it will
use this parameter and ignore the one in Z.}

\item{plan}{An audit.plan object that the audit was conducted
under.}

\item{...}{Passed to CAST.calc.sample if plan is null and needs to
be regenerated.}
}
\description{
Given audit data, compute p.values and all that.
}
