% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM_package.R
\name{KM.audit}
\alias{KM.audit}
\title{KM Audit Calculator}
\usage{
KM.audit(
  data,
  U,
  Z,
  alpha = 0.25,
  plot = FALSE,
  debug = FALSE,
  return.Ps = FALSE,
  truncate.Ps = TRUE
)
}
\arguments{
\item{data}{Data frame holding audit data with taint and tot.votes as two
columns.}

\item{U}{Maximum total error bound (sum of e.max for all batches in race).}

\item{Z}{elec.data object for the race---the original reported results.}

\item{alpha}{Risk.}

\item{plot}{Plot the audit?}

\item{debug}{Print debugging info}

\item{return.Ps}{Return the stepwise P-values}

\item{truncate.Ps}{Return the stepwise P-values only up to the audit stop
point.}
}
\value{
List of various things, including final p-value.
}
\description{
Do a KM audit given a specified list of audited batches for a specified
election.
}
\details{
This will do a single-stage KM audit as a consequence of doing the stepwise
version (since the single-stage is the same as the stepwise up to the number
of batches audited).

WARNING: This function is not fully debugged!
}
\references{
Stark, Miratrix
}
\author{
Miratrix
}
