% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAST.R
\name{make.truth.opt.bad}
\alias{make.truth.opt.bad}
\title{Given a collection of counted votes, make a theoretical bad truth by packing all error 
possible in the largest precincts}
\usage{
make.truth.opt.bad(Z, strata = "strata", bound = c("margin", "WPM"), t = 0)
}
\description{
Warning: if bound is WPM this error is made by simply adding the max amount of error
to the first loser's total (so that total votes may in this case exceed the total votes
of the precinct)--this could potentially cause trouble.  Be careful!
}
\details{
Param t: an allowed backgound level of error for all precincts
      bound.function: Can pass fractionOfVotesBound if you want to do 0.4WPM.
}
