% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAST.R
\name{make.truth.opt.bad.strat}
\alias{make.truth.opt.bad.strat}
\title{Make optimally bad truth}
\usage{
make.truth.opt.bad.strat(Z, strata = "strata", t = 3, shuffle.strata = FALSE)
}
\arguments{
\item{Z}{The elec.data to build from.}

\item{strata}{name of column holding strata, if any.}

\item{t}{Number of per-precinct vote "background" error that can occur
without triggering escalation if seen.}

\item{shuffle.strata}{Should the error be randomly put in the strata?}
}
\description{
make bad truth as described in Stark's paper (assuming fixed precinct size)
}
\seealso{
Other make.truth: 
\code{\link{make.truth.ex.bad}()}
}
\concept{make.truth}
