% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elec_functions.R
\name{maximumMarginBound}
\alias{maximumMarginBound}
\title{Election Audit Error Bound Functions}
\usage{
maximumMarginBound(Z, votes = NULL)
}
\arguments{
\item{Z}{The elec.data object.}

\item{votes}{The data.frame to compute the maximumMarginBounds for.
If null, will return all bounds for all precincts in Z.}
}
\value{
Vector (of length of precincts) of maximum possible error
  for each precinct.
}
\description{
This is one of the various bounding functions used to
bound the maximum amount of error one could see in a single audit
unit.

maximumMarginBound returns the maximum margin reduction for each
precint by computing all margin reductions between pairs of winners
& losers and then scaling by that pair's total margin to get a
proportion and then taking the max of all such proportions (usually
will be the last winner to the closest loser).
}
\author{
Luke W. Miratrix
}
