% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM_package.R
\name{print.audit.plan.KM}
\alias{print.audit.plan.KM}
\title{Pretty print KM audit plan}
\usage{
\method{print}{audit.plan.KM}(x, ...)
}
\arguments{
\item{x}{A audit.plan.KM object, such as one returned by KM.calc.sample.}

\item{...}{ignored}
}
\description{
Pretty print KM audit plan
}
