% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinomial.R
\name{tri.sample.stats}
\alias{tri.sample.stats}
\title{Utility function for tri.sample}
\usage{
tri.sample.stats(samp)
}
\arguments{
\item{samp}{A sample, such as one returned from tri.sample}
}
\value{
the total number of unique precincts and ballots given a
  sample.
}
\description{
A utility function returning the total number of unique precincts
and ballots given a sample.
}
