% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM_package.R
\name{truth.looker}
\alias{truth.looker}
\title{Looking at fake ``truths'' for election simulations}
\usage{
truth.looker(data)
}
\arguments{
\item{data}{The data.frame returned from such things as make.audit}
}
\value{
None.  Just does printout.
}
\description{
This prints out total error in a fake truth for an election, and some other
info.
}
\details{
Utility function for debugging and understanding stuff.

Look at a specific "truth" and print out what total error, etc.  is.
}
