% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elec_functions.R
\name{find.stark.SRS.p}
\alias{find.stark.SRS.p}
\title{find.stark.SRS.p}
\usage{
find.stark.SRS.p(N, n, q)
}
\arguments{
\item{N}{total number of precints}

\item{n}{total number of audited precints (must be less than N)}

\item{q}{min number of precints that could hold taint to flip election}
}
\value{
Chance that 1 or more of the q 'bad' things will be seen in a size n
SRS draw from the N sized bucket.
}
\description{
Find the p-value for a given q, n, and N.  Helper function for a simple
hypergeometric calculaton--see reports.
}
\author{
Luke W. Miratrix
}
