% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elec_data.R
\name{is.elec.data}
\alias{is.elec.data}
\title{Check if object is elec.data object}
\usage{
is.elec.data(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
is.elec.data: TRUE if object is an elec.data object.
}
\description{
Check if object is elec.data object
}
