% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM_package.R
\name{opt.sample.size}
\alias{opt.sample.size}
\title{KM Audit Sample Size Calc}
\usage{
opt.sample.size(Z, beta = 0.25)
}
\arguments{
\item{Z}{elec.data object}

\item{beta}{risk}
}
\value{
Single number of batches to sample.
}
\description{
Calc KM Optimal Sample Size
}
\details{
This is how many steps would be needed if no error was found with each step.
Obviously a bit idealistic, but still useful.
}
